+++ sim/util.hpp
@@ -3,4 +3,4 @@
 
 int getBits(int num, int start, int end, bool sext);
 void print_instruction(unsigned int instr);
-void get_filesize(std::fstream& f, int& fsize);
+void get_filesize(std::fstream& f, unsigned long long& fsize);
